// apps/web/src/api/model/getActivityLogs200ActivityLogsItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Rootly API
 * OpenAPI spec version: 0.1.0
 */
import type { GetActivityLogs200ActivityLogsItemAction } from './getActivityLogs200ActivityLogsItemAction';
import type { GetActivityLogs200ActivityLogsItemMetadata } from './getActivityLogs200ActivityLogsItemMetadata';
import type { GetActivityLogs200ActivityLogsItemResourceType } from './getActivityLogs200ActivityLogsItemResourceType';

export type GetActivityLogs200ActivityLogsItem = {
  id: string;
  workspaceId: string;
  resourceType: GetActivityLogs200ActivityLogsItemResourceType;
  resourceId: string;
  resourceName: string;
  action: GetActivityLogs200ActivityLogsItemAction;
  actorUserId: string;
  actorName: string;
  /** @nullable */
  metadata: GetActivityLogs200ActivityLogsItemMetadata;
  createdAt: string;
};